feat: use only one RequestItem for form fields
