+++ packages/consumption/src/consumption/ConsumptionController.ts
@@ -3,15 +3,14 @@ import {
     ConsentRequestItem,
     CreateAttributeRequestItem,
     DeleteAttributeRequestItem,
+    FormFieldRequestItem,
     FreeTextRequestItem,
-    FreeValueFormRequestItem,
     OwnSharedAttributeDeletedByOwnerNotificationItem,
     PeerSharedAttributeDeletedByPeerNotificationItem,
     PeerSharedAttributeSucceededNotificationItem,
     ProposeAttributeRequestItem,
     ReadAttributeRequestItem,
     RegisterAttributeListenerRequestItem,
-    SelectionFormRequestItem,
     ShareAttributeRequestItem,
     ThirdPartyRelationshipAttributeDeletedByPeerNotificationItem,
     TransferFileOwnershipRequestItem
@@ -24,8 +23,8 @@ import {
     CreateAttributeRequestItemProcessor,
     DeleteAttributeRequestItemProcessor,
     DraftsController,
+    FormFieldRequestItemProcessor,
     FreeTextRequestItemProcessor,
-    FreeValueFormRequestItemProcessor,
     GenericRequestItemProcessor,
     IdentityMetadataController,
     IncomingRequestsController,
@@ -43,7 +42,6 @@ import {
     RequestItemConstructor,
     RequestItemProcessorConstructor,
     RequestItemProcessorRegistry,
-    SelectionFormRequestItemProcessor,
     SettingsController,
     ShareAttributeRequestItemProcessor,
     ThirdPartyRelationshipAttributeDeletedByPeerNotificationItemProcessor,
@@ -165,8 +163,7 @@ export class ConsumptionController {
             [AuthenticationRequestItem, GenericRequestItemProcessor],
             [RegisterAttributeListenerRequestItem, RegisterAttributeListenerRequestItemProcessor],
             [FreeTextRequestItem, FreeTextRequestItemProcessor],
-            [FreeValueFormRequestItem, FreeValueFormRequestItemProcessor],
-            [SelectionFormRequestItem, SelectionFormRequestItemProcessor],
+            [FormFieldRequestItem, FormFieldRequestItemProcessor],
             [TransferFileOwnershipRequestItem, TransferFileOwnershipRequestItemProcessor]
         ]);
     }
