+++ packages/consumption/src/modules/requests/index.ts
@@ -14,10 +14,10 @@ export * from "./itemProcessors/AbstractRequestItemProcessor";
 export * from "./itemProcessors/createAttribute/CreateAttributeRequestItemProcessor";
 export * from "./itemProcessors/deleteAttribute/AcceptDeleteAttributeRequestItemParameters";
 export * from "./itemProcessors/deleteAttribute/DeleteAttributeRequestItemProcessor";
+export * from "./itemProcessors/formField/AcceptFormFieldRequestItemParameters";
+export * from "./itemProcessors/formField/FormFieldRequestItemProcessor";
 export * from "./itemProcessors/freeText/AcceptFreeTextRequestItemParameters";
 export * from "./itemProcessors/freeText/FreeTextRequestItemProcessor";
-export * from "./itemProcessors/freeValueForm/AcceptFreeValueFormRequestItemParameters";
-export * from "./itemProcessors/freeValueForm/FreeValueFormRequestItemProcessor";
 export * from "./itemProcessors/GenericRequestItemProcessor";
 export * from "./itemProcessors/IRequestItemProcessor";
 export * from "./itemProcessors/proposeAttribute/AcceptProposeAttributeRequestItemParameters";
@@ -32,8 +32,6 @@ export * from "./itemProcessors/registerAttributeListener/RegisterAttributeListe
 export * from "./itemProcessors/RequestItemConstructor";
 export * from "./itemProcessors/RequestItemProcessorConstructor";
 export * from "./itemProcessors/RequestItemProcessorRegistry";
-export * from "./itemProcessors/selectionForm/AcceptSelectionFormRequestItemParameters";
-export * from "./itemProcessors/selectionForm/SelectionFormRequestItemProcessor";
 export * from "./itemProcessors/shareAttribute/ShareAttributeRequestItemProcessor";
 export * from "./itemProcessors/transferFileOwnership/TransferFileOwnershipRequestItemProcessor";
 export * from "./local/LocalRequest";
