+++ packages/runtime/src/dataViews/DataViewExpander.ts
@@ -11,10 +11,10 @@ import {
     DeleteAttributeRequestItemJSON,
     DisplayNameJSON,
     ErrorResponseItemJSON,
+    FormFieldAcceptResponseItemJSON,
+    FormFieldRequestItemJSON,
     FreeTextAcceptResponseItemJSON,
     FreeTextRequestItemJSON,
-    FreeValueFormAcceptResponseItemJSON,
-    FreeValueFormRequestItemJSON,
     GivenNameJSON,
     IQLQueryJSON,
     IdentityAttribute,
@@ -44,7 +44,6 @@ import {
     ResponseItemJSON,
     ResponseItemResult,
     ResponseJSON,
-    SelectionFormRequestItemJSON,
     SexJSON,
     ShareAttributeAcceptResponseItemJSON,
     ShareAttributeRequestItemJSON,
@@ -56,7 +55,6 @@ import {
     ValueHintsJSON,
     isRequestItemDerivation
 } from "@nmshd/content";
-import { SelectionFormAcceptResponseItemJSON } from "@nmshd/content/src/requests/items/selectionForm/SelectionFormAcceptResponseItem";
 import { CoreAddress, CoreId, FileReference } from "@nmshd/core-types";
 import { IdentityController } from "@nmshd/transport";
 import { Inject } from "@nmshd/typescript-ioc";
@@ -85,12 +83,11 @@ import {
     DecidableConsentRequestItemDVO,
     DecidableCreateAttributeRequestItemDVO,
     DecidableDeleteAttributeRequestItemDVO,
+    DecidableFormFieldRequestItemDVO,
     DecidableFreeTextRequestItemDVO,
-    DecidableFreeValueFormRequestItemDVO,
     DecidableProposeAttributeRequestItemDVO,
     DecidableReadAttributeRequestItemDVO,
     DecidableRegisterAttributeListenerRequestItemDVO,
-    DecidableSelectionFormRequestItemDVO,
     DecidableShareAttributeRequestItemDVO,
     DecidableTransferFileOwnershipRequestItemDVO,
     LocalAttributeDVO,
@@ -123,10 +120,10 @@ import {
     DraftIdentityAttributeDVO,
     DraftRelationshipAttributeDVO,
     ErrorResponseItemDVO,
+    FormFieldAcceptResponseItemDVO,
+    FormFieldRequestItemDVO,
     FreeTextAcceptResponseItemDVO,
     FreeTextRequestItemDVO,
-    FreeValueFormAcceptResponseItemDVO,
-    FreeValueFormRequestItemDVO,
     IQLQueryDVO,
     IdentityAttributeQueryDVO,
     MailDVO,
@@ -146,8 +143,6 @@ import {
     ResponseDVO,
     ResponseItemDVO,
     ResponseItemGroupDVO,
-    SelectionFormAcceptResponseItemDVO,
-    SelectionFormRequestItemDVO,
     ShareAttributeAcceptResponseItemDVO,
     ShareAttributeRequestItemDVO,
     ThirdPartyRelationshipAttributeQueryDVO,
@@ -730,27 +725,27 @@ export class DataViewExpander {
                     response: responseItemDVO
                 } as FreeTextRequestItemDVO;
 
-            case "FreeValueFormRequestItem":
-                const freeValueFormRequestItem = requestItem as FreeValueFormRequestItemJSON;
+            case "FormFieldRequestItem":
+                const formFieldRequestItem = requestItem as FormFieldRequestItemJSON;
 
                 if (isDecidable) {
                     return {
-                        ...freeValueFormRequestItem,
-                        type: "DecidableFreeValueFormRequestItemDVO",
+                        ...formFieldRequestItem,
+                        type: "DecidableFormFieldRequestItemDVO",
                         id: "",
-                        name: requestItem.title ?? "i18n://dvo.requestItem.DecidableFreeValueFormRequestItem.name",
+                        name: requestItem.title ?? "i18n://dvo.requestItem.DecidableFormFieldRequestItem.name",
                         isDecidable,
                         response: responseItemDVO
-                    } as DecidableFreeValueFormRequestItemDVO;
+                    } as DecidableFormFieldRequestItemDVO;
                 }
                 return {
-                    ...freeValueFormRequestItem,
-                    type: "FreeValueFormRequestItemDVO",
+                    ...formFieldRequestItem,
+                    type: "FormFieldRequestItemDVO",
                     id: "",
-                    name: requestItem.title ?? "i18n://dvo.requestItem.FreeValueFormRequestItem.name",
+                    name: requestItem.title ?? "i18n://dvo.requestItem.FormFieldRequestItem.name",
                     isDecidable,
                     response: responseItemDVO
-                } as FreeValueFormRequestItemDVO;
+                } as FormFieldRequestItemDVO;
 
             case "RegisterAttributeListenerRequestItem":
                 const registerAttributeListenerRequestItem = requestItem as RegisterAttributeListenerRequestItemJSON;
@@ -778,28 +773,6 @@ export class DataViewExpander {
                     isDecidable,
                     response: responseItemDVO
                 } as RegisterAttributeListenerRequestItemDVO;
-
-            case "SelectionFormRequestItem":
-                const selectionFormRequestItem = requestItem as SelectionFormRequestItemJSON;
-
-                if (isDecidable) {
-                    return {
-                        ...selectionFormRequestItem,
-                        type: "DecidableSelectionFormRequestItemDVO",
-                        id: "",
-                        name: requestItem.title ?? "i18n://dvo.requestItem.DecidableSelectionFormRequestItem.name",
-                        isDecidable,
-                        response: responseItemDVO
-                    } as DecidableSelectionFormRequestItemDVO;
-                }
-                return {
-                    ...selectionFormRequestItem,
-                    type: "SelectionFormRequestItemDVO",
-                    id: "",
-                    name: requestItem.title ?? "i18n://dvo.requestItem.SelectionFormRequestItem.name",
-                    isDecidable,
-                    response: responseItemDVO
-                } as SelectionFormRequestItemDVO;
 
             case "TransferFileOwnershipRequestItem":
                 const transferFileOwnershipRequestItem = requestItem as TransferFileOwnershipRequestItemJSON;
@@ -954,15 +927,15 @@ export class DataViewExpander {
                         name: name
                     } as FreeTextAcceptResponseItemDVO;
 
-                case "FreeValueFormAcceptResponseItem":
-                    const freeValueFormResponseItem = responseItem as FreeValueFormAcceptResponseItemJSON;
+                case "FormFieldAcceptResponseItem":
+                    const freeValueFormResponseItem = responseItem as FormFieldAcceptResponseItemJSON;
 
                     return {
                         ...freeValueFormResponseItem,
-                        type: "FreeValueFormAcceptResponseItemDVO",
+                        type: "FormFieldAcceptResponseItemDVO",
                         id: "",
                         name: name
-                    } as FreeValueFormAcceptResponseItemDVO;
+                    } as FormFieldAcceptResponseItemDVO;
 
                 case "RegisterAttributeListenerAcceptResponseItem":
                     const registerAttributeListenerResponseItem = responseItem as RegisterAttributeListenerAcceptResponseItemJSON;
@@ -976,16 +949,6 @@ export class DataViewExpander {
                         name: name,
                         listener: localAttributeListener
                     } as RegisterAttributeListenerAcceptResponseItemDVO;
-
-                case "SelectionFormAcceptResponseItem":
-                    const selectionFormResponseItem = responseItem as SelectionFormAcceptResponseItemJSON;
-
-                    return {
-                        ...selectionFormResponseItem,
-                        type: "SelectionFormAcceptResponseItemDVO",
-                        id: "",
-                        name: name
-                    } as SelectionFormAcceptResponseItemDVO;
 
                 case "TransferFileOwnershipAcceptResponseItem":
                     const transferFileOwnershipResponseItem = responseItem as TransferFileOwnershipAcceptResponseItemJSON;
